requirements.txt
psycopg2-binary==2.8.6
SQLAlchemy==1.3.20
PyYAML==5.3.1
pytest==6.1.1
